requirements.txt
asgiref==3.8.1
Django==4.2.18
djangorestframework==3.15.2
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.3
typing-extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
